--- my-electron-react-app/src/Componentes/cronometro/cronometro.css
@@ -1,82 +1,0 @@
-.cronometro-container {
-    font-family: 'Arial', sans-serif;
-    max-width: 400px;
-    margin: 20px auto;
-    padding: 20px;
-    background-color: #f5f5f5;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .display {
-    font-size: 2.5rem;
-    font-weight: bold;
-    text-align: center;
-    margin: 20px 0;
-    color: #333;
-    background-color: #e0e0e0;
-    padding: 15px;
-    border-radius: 5px;
-    font-family: 'Segment7Standard', monospace;
-  }
-  
-  .milisegundos {
-    font-size: 1.5rem;
-    color: #666;
-  }
-  
-  .controles {
-    display: flex;
-    justify-content: center;
-    gap: 10px;
-    margin-bottom: 20px;
-  }
-  
-  button {
-    padding: 8px 15px;
-    border: none;
-    border-radius: 5px;
-    background-color: #4CAF50;
-    color: white;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  button:hover {
-    background-color: #45a049;
-  }
-  
-  button:disabled {
-    background-color: #cccccc;
-    cursor: not-allowed;
-  }
-  
-  .vueltas-container {
-    margin-top: 20px;
-    padding: 15px;
-    background-color: #e9e9e9;
-    border-radius: 5px;
-  }
-  
-  .vueltas-container h3 {
-    margin-top: 0;
-    color: #333;
-  }
-  
-  .vueltas-container ul {
-    list-style-type: none;
-    padding: 0;
-  }
-  
-  .vueltas-container li {
-    padding: 5px 0;
-    border-bottom: 1px solid #ddd;
-  }
-  
-  /* Fuente para display digital (opcional) */
-  @font-face {
-    font-family: 'Segment7Standard';
-    src: url('Segment7Standard.otf') format('opentype');
-    font-weight: normal;
-    font-style: italic;
-  }
